front: move followup components to dedicated folder (#911)
