+++ public/mobile-app/src/lib/components/followup/FollowupInformation.svelte
