+++ public/mobile-app/src/lib/components/followup/FollowupItem.svelte.test.ts
