+++ public/mobile-app/src/lib/components/followup/FollowupItemDetailHeader.svelte
