+++ public/mobile-app/src/routes/followup/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { onMount } from 'svelte';
   import { goto } from '$app/navigation';
-  import Followup from '$lib/components/Followup.svelte';
+  import Followup from '$lib/components/followup/Followup.svelte';
   import Navigation from '$lib/components/Navigation.svelte';
   import { userStore } from '$lib/state/User.svelte';
