+++ public/mobile-app/src/routes/followup/archived/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { onMount } from 'svelte';
   import { goto } from '$app/navigation';
-  import Followup from '$lib/components/Followup.svelte';
+  import Followup from '$lib/components/followup/Followup.svelte';
   import { userStore } from '$lib/state/User.svelte';
 
   onMount(async () => {
